requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-telegram-bot==13.15
schedule==1.2.0
python-dotenv==1.0.0
apscheduler==3.6.3
httpx==0.24.1
lxml==4.9.3